Moved the company-mapping service
I will also need it for import/export
